Ignore .clang_format file not found (#26)
Adding argument to the clang-format tool plugin to allow for not raising an exception when the configuration file does not match. This is motivated by that feature not working (so far) in Travis CI jobs. Travis configuration files can now be setup to ignore the exception while still leaving the Statick configuration such that clang-format works as expected on CLI and Jenkins.