Fixed the uncrustify unit test so it works on Ubuntu Bionic. (#160)
The version of uncrustify on Xenial is 0.59 and on Bionic is 0.66.
Between release of those versions the option align_number_left was
changed to align_number_right. When uncrustify is run as part of the
unit test in Bionic the tool fails to parse the configuration file
correctly, thus causing the unit test to fail on Bionic.

I copied the original configuration file, removed that line, and placed
it in a path relative to the unit test. I setup the unit test to use
this new path as part of the user_paths resource. This causes the
uncrustify tool to use this new configuration file when running the tool
as part of the unit test.